/input/{sctp,tcp,udp}/{sflow.c,sflow.h}: Fixed unsigned comparison against zero (102097)

+++ base/src/input/tcp/sflow.c
@@ -60,7 +60,7 @@ extern "C" {
 #define YES 1
 #define NO 0
 
-static int numOfFlowSamples;
+static uint16_t numOfFlowSamples;
 
 /* define my own IP header struct - to ease portability */
 struct myiphdr
@@ -1843,7 +1843,7 @@ static void readSFlowDatagram(SFSample *sample, char *packet)
   }
 }
 
-int Process_sflow(void *packet, ssize_t packet_len) {
+uint16_t Process_sflow(void *packet, ssize_t packet_len) {
 
 SFSample 	sample;
 int 		exceptionVal;
